refactor: Update navigation options for archived screen
