+++ app/(pages)/_layout.tsx
@@ -80,6 +80,41 @@ export default function ModalLayout() {
           },
         }}
       />
+      <Stack.Screen
+        name="archive"
+        options={{
+          title: "Archived Chats",
+          headerShown: true,
+          headerBackVisible: true,
+          headerLargeTitle: true,
+          headerShadowVisible: false,
+          headerBackTitleVisible: true,
+          headerBackButtonMenuEnabled: true,
+          headerBackTitle: "Back",
+          headerTintColor:
+            colorScheme === "dark" ? Colors.dark.black : Colors.light.black,
+          headerStyle: {
+            backgroundColor:
+              colorScheme === "dark"
+                ? Colors.dark.background
+                : Colors.light.background,
+          },
+          headerLargeTitleStyle: {
+            color:
+              colorScheme === "dark" ? Colors.dark.black : Colors.light.black,
+            fontFamily: "Lexend-Bold",
+          },
+          headerBackTitleStyle: {
+            fontFamily: "Lexend-Bold",
+          },
+          headerTitleStyle: {
+            fontSize: 18,
+            fontFamily: "Lexend-Bold",
+            color:
+              colorScheme === "dark" ? Colors.dark.black : Colors.light.black,
+          },
+        }}
+      />
     </Stack>
   );
 }
