--- app/(tabs)/rooms/archive.tsx
@@ -1,12 +1,0 @@
-import { View, Text } from "react-native";
-import React from "react";
-
-const ArchivedScreen = () => {
-  return (
-    <View>
-      <Text>ArchivedScreen</Text>
-    </View>
-  );
-};
-
-export default ArchivedScreen;
